+++ src/Libraries/SmartStore.Core/Domain/Media/MediaFile.cs
@@ -18,6 +18,13 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets the associated folder identifier.
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_MediaType", 0)]
+		[Index("IX_Media_Extension", 0)]
+		[Index("IX_Media_PixelSize", 0)]
+		[Index("IX_Media_Name", 0)]
+		[Index("IX_Media_Size", 0)]
+		[Index("IX_Media_UpdatedOnUtc", 0)]
+		[Index("IX_Media_FolderId", 0)]
 		public int? FolderId { get; set; }
 
         /// <summary>
@@ -30,6 +37,7 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets the SEO friendly name of the media file including file extension
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_Name", 1)]
 		public string Name { get; set; }
 
 		/// <summary>
@@ -48,6 +56,8 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets the (dotless) file extension
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_MediaType", 2)]
+		[Index("IX_Media_Extension", 1)]
 		public string Extension { get; set; }
 
 		/// <summary>
@@ -60,18 +70,23 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets the file media type (image, video, audio, document etc.)
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_MediaType", 1)]
 		public string MediaType { get; set; }
 
 		/// <summary>
 		/// Gets or sets the file size in bytes
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_Size", 1)]
 		public int Size { get; set; }
 
 		/// <summary>
 		/// Gets or sets the total pixel size of an image (width * height)
 		/// </summary>
 		[DataMember]
+		[Index("IX_Media_MediaType", 3)]
+		[Index("IX_Media_Extension", 2)]
+		[Index("IX_Media_PixelSize", 1)]
 		public int? PixelSize { get; set; }
 
 		/// <summary>
@@ -96,7 +111,6 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets the date and time of instance update
 		/// </summary>
 		[DataMember]
-		[Index("IX_CreatedOn_IsTransient", 0)]
 		public DateTime CreatedOnUtc { get; set; }
 
 		/// <summary>
@@ -109,12 +123,18 @@ namespace SmartStore.Core.Domain.Media
 		/// Gets or sets a value indicating whether the file is transient/preliminary
 		/// </summary>
 		[DataMember]
-		[Index("IX_CreatedOn_IsTransient", 1)]
 		public bool IsTransient { get; set; }
 
 		/// <summary>
 		/// Gets or sets a value indicating whether the file has been soft deleted
 		/// </summary>
+		[Index("IX_Media_MediaType", 4)]
+		[Index("IX_Media_Extension", 3)]
+		[Index("IX_Media_PixelSize", 2)]
+		[Index("IX_Media_Name", 2)]
+		[Index("IX_Media_Size", 2)]
+		[Index("IX_Media_UpdatedOnUtc", 2)]
+		[Index("IX_Media_FolderId", 1)]
 		public bool Deleted { get; set; }
 
 		/// <summary>