+++ src/Libraries/SmartStore.Services/Cms/Blocks/IBlockHandler.cs
@@ -28,11 +28,25 @@ namespace SmartStore.Services.Cms.Blocks
 		Edit
 	}
 
+	public interface IBlockHtmlParts
+	{
+		void AddCssClass(string value);
+		void AddHtmlAttribute(string name, object value);
+	}
+
 	/// <summary>
 	/// Handles rendering of corresponding <see cref="IBlock"/> implementations.
 	/// </summary>
 	public interface IBlockHandler
 	{
+		/// <summary>
+		/// Called when the block is about to be rendered. Gives you the chance to add css classes or html attributes to the root element.
+		/// </summary>
+		/// <param name="container">The block instance wrapper / model.</param>
+		/// <param name="viewMode">The stories current view mode.</param>
+		/// <param name="htmlParts">Allows adding classes or attributes.</param>
+		void BeforeRender(IBlockContainer container, StoryViewMode viewMode, IBlockHtmlParts htmlParts);
+
 		/// <summary>
 		/// Renders the block result.
 		/// </summary>
@@ -40,6 +54,24 @@ namespace SmartStore.Services.Cms.Blocks
 		/// <param name="templates">A list of template names. The first valid template will be used for rendering.</param>
 		/// <param name="htmlHeper">Html helper instance.</param>
 		void Render(IBlockContainer element, IEnumerable<string> templates, HtmlHelper htmlHeper);
+
+		/// <summary>
+		/// Called after the entity has been saved to the database to perform operations which require an entity id (e.g. localization stuff).
+		/// </summary>
+		/// <param name="container">The block instance wrapper / model.</param>
+		/// <param name="entity">
+		/// The corresponding entity record for the block.
+		/// </param>
+		void AfterSave(IBlockContainer container, IBlockEntity entity);
+
+		/// <summary>
+		/// Clones <see cref="IBlockEntity.Model"/>. In most cases it is sufficient to directly return the serialized model, but there may be cases
+		/// where the inner data must be cloned first, e.g. if it contains a picture reference.
+		/// </summary>
+		/// <param name="sourceEntity">The source entity.</param>
+		/// <param name="clonedEntity">The target entity clone which is about to be saved in the data storage.</param>
+		/// <returns>The serialized block clone.</returns>
+		string Clone(IBlockEntity sourceEntity, IBlockEntity clonedEntity);
 
 		/// <summary>
 		/// returns the block's render result as string.
@@ -81,23 +113,5 @@ namespace SmartStore.Services.Cms.Blocks
         /// be converted to JSON and assigned to <see cref="IBlockEntity.Model"/> property, which then will be saved in the data storage.
         /// </param>
         void Save(T block, IBlockEntity entity);
-
-        /// <summary>
-        /// Called after the entity has been saved to the database to perform operations which require an entity id (e.g. localization stuff).
-        /// </summary>
-        /// <param name="container">The block instance wrapper / model.</param>
-        /// <param name="entity">
-        /// The corresponding entity record for the block.
-        /// </param>
-        void AfterSave(IBlockContainer container, IBlockEntity entity);
-
-        /// <summary>
-        /// Clones <see cref="IBlockEntity.Model"/>. In most cases it is sufficient to directly return the serialized model, but there may be cases
-        /// where the inner data must be cloned first, e.g. if it contains a picture reference.
-        /// </summary>
-        /// <param name="sourceEntity">The source entity.</param>
-        /// <param name="clonedEntity">The target entity clone which is about to be saved in the data storage.</param>
-        /// <returns>The serialized block clone.</returns>
-        string Clone(IBlockEntity sourceEntity, IBlockEntity clonedEntity);
 	}
 }