+++ src/Libraries/SmartStore.Services/Media/Migration/MediaMigrator.cs
@@ -17,11 +17,22 @@ using System.Runtime.CompilerServices;
 using SmartStore.Core.Domain.Messages;
 using System.Diagnostics;
 using System.Globalization;
+using NReco.PdfGenerator;
+using SmartStore.Core;
 
 namespace SmartStore.Services.Media.Migration
 {
     public class MediaMigrator
     {
+        public class FileMin
+        {
+            public int Id { get; set; }
+            public string Name { get; set; }
+            public bool IsDupe { get; set; }
+            public string UniqueName { get; set; }
+            public int UniqueIndex { get; set; }
+        }
+        
         internal static bool Executed;
         internal const string MigrationName = "MediaManager";
         
@@ -78,7 +89,6 @@ namespace SmartStore.Services.Media.Migration
                 Execute("MigrateMediaFiles", () => MigrateMediaFiles(ctx));
                 Execute("MigrateUploadedFiles", () => MigrateUploadedFiles(ctx));
                 Execute("DetectTracks", () => DetectTracks());
-                Execute("DetectDupeFiles", () => DetectDupeFiles(ctx));
                 Execute("ClearImageCache", () => ClearImageCache());
 
                 _folderService.ClearCache();
@@ -311,6 +321,82 @@ namespace SmartStore.Services.Media.Migration
 
         public void MigrateMediaFiles(SmartObjectContext ctx)
         {
+            string prevName = null;
+            int fileIndex = 0;
+
+            var query = ctx.Set<MediaFile>().OrderBy(x => x.Name);
+            var totalCount = query.Count();
+            var pageSize = Math.Max(1000, Math.Min(5000, totalCount / 200));
+            var pageIndex = 0;
+
+            using (var scope = new DbContextScope(ctx,
+                hooksEnabled: false,
+                autoCommit: false,
+                proxyCreation: false,
+                validateOnSave: false,
+                lazyLoading: false))
+            {
+                while (true)
+                {
+                    var files = new PagedList<MediaFile>(query, pageIndex, pageSize);
+                    if (files.Count == 0)
+                        break;
+
+                    foreach (var file in files)
+                    {
+                        if (file.Version > 0)
+                            continue;
+
+                        if (file.Extension.IsEmpty())
+                        {
+                            file.Extension = MimeTypes.MapMimeTypeToExtension(file.MimeType);
+                        }
+
+                        var name = file.Name;
+                        var fixedName = name;
+                        if (name.IsEmpty())
+                        {
+                            name = fixedName = file.Id.ToString(System.Globalization.CultureInfo.InvariantCulture);
+                        }
+                        else
+                        {
+                            name = fixedName = file.Name.Truncate(292).ToValidFileName();
+                            if (name == prevName)
+                            {
+                                // Make file name unique
+                                fixedName = name + "-" + ++fileIndex; 
+                            }
+                            else
+                            {
+                                fileIndex = 0;
+                            }
+                        }
+
+                        prevName = name;
+
+                        file.Name = fixedName + "." + file.Extension;
+                        file.CreatedOnUtc = file.UpdatedOnUtc;
+                        file.Version = 1;
+
+                        ProcessMediaFile(file);
+                    }
+
+                    // Save to DB
+                    int num = scope.Commit();
+
+                    // Breathe
+                    ctx.DetachEntities<MediaFile>(deep: true);
+
+                    if (!files.HasNextPage)
+                        break;
+
+                    pageIndex++;
+                }
+            }
+        }
+
+        public void MigrateMediaFiles_Old(SmartObjectContext ctx)
+        {
             var query = ctx.Set<MediaFile>();
                 //.Where(x => x.Version == 0)
                 //.Include(x => x.MediaStorage);
@@ -517,20 +603,6 @@ namespace SmartStore.Services.Media.Migration
                 
                 _mediaTracker.DetectAllTracks(albumName, true);
             }
-        }
-
-        public void DetectDupeFiles(SmartObjectContext ctx)
-        {
-            //var lookup = ctx.Set<MediaFile>()
-            //            .Select(f => new { f.Id, f.FolderId, f.Name })
-            //            .GroupBy(f => f.FolderId)
-            //            .ToList();
-
-            //foreach (var folder in lookup)
-            //{
-            //    var allFileNames = new HashSet<string>(folder.Select(x => x.Name), StringComparer.OrdinalIgnoreCase);
-            //    //folder.
-            //}
         }
 
         public void ClearImageCache()