+++ src/Presentation/SmartStore.Web/Content/editors/summernote/plugins/smartstore.link.js
@@ -309,23 +309,23 @@
 				var toggleLinkBtn = function () {
 					var enable = $linkUrl.val();
 					if (!linkInfo.img) {
-						enable = enable && $linkText.val();
+						enable = !!(enable) && !!($linkText.val());
 					}
 					ui.toggleBtn($linkBtn, enable);
 				};
 				var handleLinkTextUpdate = function () {
-					toggleLinkBtn();
 					// if linktext was modified by keyup,
 					// stop cloning text from linkUrl
 					linkInfo.text = $linkText.val();
+					toggleLinkBtn();
 				};
 				var handleLinkUrlUpdate = function () {
-					toggleLinkBtn();
 					// display same link on `Text to display` input
 					// when create a new link
 					if (!linkInfo.text) {
 						$linkText.val($linkUrl.val());
 					}
+					toggleLinkBtn();
 				};
 
 				$linkText.on('input.linkDialog', handleLinkTextUpdate);